refactor(healthplatform): remove issuereporter shim, use store.Component directly
The issuereporter interface was a 2-method subset of store.Component with no
consumers outside this PR. Checks implementing IssueAwareCheck now take
store.Component directly, removing an unnecessary indirection layer.

- Delete comp/healthplatform/issuereporter/def/
- Remove from tasks/components.py allowlist and comp/README.md
- Update IssueAwareCheck.SetIssueReporter, NewCheckWrapper, and call sites
  (collector.go, collector_test.go, check_wrapper_test.go) to use store.Component
- Update pkg/collector/check/BUILD.bazel dep accordingly
