+++ comp/collector/collector/impl/internal/middleware/check_wrapper_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 
 	agenttelemetry "github.com/DataDog/datadog-agent/comp/core/agenttelemetry/def"
-	issuereporter "github.com/DataDog/datadog-agent/comp/healthplatform/issuereporter/def"
+	healthplatformstore "github.com/DataDog/datadog-agent/comp/healthplatform/store/def"
 	"github.com/DataDog/datadog-agent/pkg/collector/check"
 	installertelemetry "github.com/DataDog/datadog-agent/pkg/fleet/installer/telemetry"
 	"github.com/DataDog/datadog-agent/pkg/util/option"
@@ -49,15 +49,15 @@ func (m *mockTelemetry) SendEvent(_ string, _ []byte) error {
 }
 
 type mockIssueReporter struct {
-	issuereporter.Component
+	healthplatformstore.Component
 }
 
 type issueAwareCheck struct {
 	mockCheck
-	reporter issuereporter.Component
+	reporter healthplatformstore.Component
 }
 
-func (c *issueAwareCheck) SetIssueReporter(r issuereporter.Component) {
+func (c *issueAwareCheck) SetIssueReporter(r healthplatformstore.Component) {
 	c.reporter = r
 }
 
@@ -69,7 +69,7 @@ func TestCheckWrapperInjectsIssueReporter(t *testing.T) {
 		inner,
 		nil,
 		option.None[agenttelemetry.Component](),
-		option.New[issuereporter.Component](reporter),
+		option.New[healthplatformstore.Component](reporter),
 	)
 
 	require.NotNil(t, wrapper)
@@ -84,7 +84,7 @@ func TestCheckWrapperSkipsNonIssueAwareCheck(t *testing.T) {
 		inner,
 		nil,
 		option.None[agenttelemetry.Component](),
-		option.New[issuereporter.Component](reporter),
+		option.New[healthplatformstore.Component](reporter),
 	)
 
 	require.NotNil(t, wrapper)
@@ -104,7 +104,7 @@ func TestCheckWrapperCreatesSpan(t *testing.T) {
 		mockCheck,
 		nil, // senderManager is not needed for this test
 		option.New[agenttelemetry.Component](mockTelemetry),
-		option.None[issuereporter.Component](),
+		option.None[healthplatformstore.Component](),
 	)
 
 	// Run the check
