--- comp/healthplatform/issuereporter/def/BUILD.bazel
@@ -1,9 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "def",
-    srcs = ["component.go"],
-    importpath = "github.com/DataDog/datadog-agent/comp/healthplatform/issuereporter/def",
-    visibility = ["//visibility:public"],
-    deps = ["@com_github_datadog_agent_payload_v5//healthplatform"],
-)
